[Refactor] 커스텀 훅과 UI 분리
